Use GitHub Actions for R CMD check.
Refer to https://github.com/r-lib/actions/tree/master/examples for
more inspiration.

Notice that for r-devel we supply custom http-user-agent
header, it seems important. Also for r-devel grepbox needs to be
installed from repo.

windows-latest is not added to test matrix, because it fails on doMC package
availability (not available for Windows).